requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt --strip-extras src/pyproject.toml
#
agate==1.9.1
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-snowflake
annotated-types==0.7.0
    # via pydantic
asn1crypto==1.5.1
    # via snowflake-connector-python
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
babel==2.16.0
    # via agate
boto3-stubs==1.35.46
    # via dbt-lambda (src/pyproject.toml)
botocore-stubs==1.35.47
    # via boto3-stubs
certifi==2024.8.30
    # via
    #   requests
    #   snowflake-connector-python
cffi==1.17.1
    # via
    #   cryptography
    #   snowflake-connector-python
charset-normalizer==3.4.0
    # via
    #   requests
    #   snowflake-connector-python
click==8.1.7
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
colorama==0.4.6
    # via dbt-common
cryptography==43.0.3
    # via
    #   pyopenssl
    #   secretstorage
    #   snowflake-connector-python
daff==1.3.46
    # via dbt-core
dbt-adapters==1.7.0
    # via
    #   dbt-core
    #   dbt-snowflake
dbt-common==1.10.0
    # via
    #   dbt-adapters
    #   dbt-core
    #   dbt-snowflake
dbt-core==1.8.8
    # via dbt-snowflake
dbt-extractor==0.5.1
    # via dbt-core
dbt-semantic-interfaces==0.5.1
    # via dbt-core
dbt-snowflake==1.8.4
    # via dbt-lambda (src/pyproject.toml)
deepdiff==7.0.1
    # via dbt-common
filelock==3.16.1
    # via snowflake-connector-python
idna==3.10
    # via
    #   requests
    #   snowflake-connector-python
importlib-metadata==6.11.0
    # via dbt-semantic-interfaces
isodate==0.6.1
    # via
    #   agate
    #   dbt-common
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.1.0
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.4
    # via
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema==4.23.0
    # via
    #   dbt-common
    #   dbt-semantic-interfaces
jsonschema-specifications==2024.10.1
    # via jsonschema
keyring==25.4.1
    # via snowflake-connector-python
leather==0.4.0
    # via agate
logbook==1.5.3
    # via dbt-core
markupsafe==3.0.2
    # via jinja2
mashumaro==3.14
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
minimal-snowplow-tracker==0.0.2
    # via dbt-core
more-itertools==10.5.0
    # via
    #   dbt-semantic-interfaces
    #   jaraco-classes
    #   jaraco-functools
msgpack==1.1.0
    # via mashumaro
mypy-boto3-codecommit==1.35.0
    # via boto3-stubs
mypy-boto3-s3==1.35.46
    # via boto3-stubs
mypy-boto3-secretsmanager==1.35.0
    # via boto3-stubs
networkx==3.4.2
    # via dbt-core
ordered-set==4.1.0
    # via deepdiff
packaging==24.1
    # via
    #   dbt-core
    #   snowflake-connector-python
parsedatetime==2.6
    # via agate
pathspec==0.12.1
    # via
    #   dbt-common
    #   dbt-core
platformdirs==4.3.6
    # via snowflake-connector-python
protobuf==4.25.5
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via dbt-semantic-interfaces
pydantic-core==2.23.4
    # via pydantic
pyjwt==2.9.0
    # via snowflake-connector-python
pyopenssl==24.2.1
    # via snowflake-connector-python
python-dateutil==2.9.0.post0
    # via
    #   dbt-common
    #   dbt-semantic-interfaces
python-slugify==8.0.4
    # via agate
pytimeparse==1.1.8
    # via agate
pytz==2024.2
    # via
    #   dbt-adapters
    #   dbt-core
    #   dbt-lambda (src/pyproject.toml)
    #   snowflake-connector-python
pyyaml==6.0.2
    # via
    #   dbt-core
    #   dbt-lambda (src/pyproject.toml)
    #   dbt-semantic-interfaces
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   dbt-common
    #   dbt-core
    #   minimal-snowplow-tracker
    #   snowflake-connector-python
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
secretstorage==3.3.3
    # via keyring
six==1.16.0
    # via
    #   isodate
    #   minimal-snowplow-tracker
    #   python-dateutil
snowflake-connector-python==3.12.2
    # via dbt-snowflake
sortedcontainers==2.4.0
    # via snowflake-connector-python
sqlparse==0.5.1
    # via dbt-core
text-unidecode==1.3
    # via python-slugify
tomlkit==0.13.2
    # via snowflake-connector-python
types-awscrt==0.23.0
    # via botocore-stubs
types-s3transfer==0.10.3
    # via boto3-stubs
typing-extensions==4.12.2
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
    #   mashumaro
    #   pydantic
    #   pydantic-core
    #   snowflake-connector-python
urllib3==2.2.3
    # via requests
zipp==3.20.2
    # via importlib-metadata
